+++ app/services/lab/lims/pull_worker.rb
@@ -204,10 +204,10 @@ module Lab
 
           creator = format_result_entered_by(test_results['result_entered_by'])
 
-          ResultsService.create_results(test.id, provider_id: User.current.person_id,
-                                                 date: Utils.parse_date(test_results['date_result_entered'], order[:order_date].to_s),
-                                                 comments: "LIMS import: Entered by: #{creator}",
-                                                 measures: measures)
+          ResultsService.create_results(test.id,  { provider_id: User.current.person_id,
+                                                    date: Utils.parse_date(test_results['date_result_entered'], order[:order_date].to_s),
+                                                    comments: "LIMS import: Entered by: #{creator}",
+                                                    measures: measures } )
         end
       end